+++ crates/yomikiri-backend/Cargo.toml
@@ -10,9 +10,6 @@ license = "MIT"
 [lib]
 crate-type = ["cdylib", "rlib", "staticlib"]
 
-[features]
-uniffi = ["dep:uniffi"]
-
 
 [dependencies]
 lindera-tokenizer = { workspace = true }
@@ -31,7 +28,6 @@ fst = { workspace = true }
 anyhow = { workspace = true }
 regex = { workspace = true }
 japanese-utils = { workspace = true }
-uniffi = { workspace = true, optional = true }
 
 [target.'cfg(target_family="wasm")'.dependencies]
 serde-wasm-bindgen = { version = "0.6" }
@@ -56,7 +52,7 @@ memmap2 = { workspace = true }
 fs-err = { workspace = true }
 
 [lints.rust]
-unexpected_cfgs = { level = "allow", check-cfg = ['cfg(uniffi)', 'cfg(wasm)'] }
+unexpected_cfgs = { level = "allow", check-cfg = ['cfg(wasm)'] }
 
 [lints.clippy]
 needless_borrows_for_generic_args = "allow"
