+++ crates/yomikiri-backend/build.rs
@@ -3,7 +3,6 @@ use cfg_aliases::cfg_aliases;
 fn main() {
     cfg_aliases! {
       wasm: { target_family="wasm" },
-      uniffi: { feature="uniffi" },
       apple: { any(target_os="macos", target_os="ios") }
     }
 }
