+++ crates/yomikiri-backend/src/error.rs
@@ -8,7 +8,6 @@ use wasm_bindgen::prelude::*;
 pub type WasmResult<T> = Result<T, BackendError>;
 
 #[derive(Debug, Serialize)]
-#[cfg_attr(uniffi, derive(uniffi::Object))]
 #[cfg_attr(wasm, wasm_bindgen(getter_with_clone))]
 pub struct BackendError {
     pub message: String,
@@ -32,44 +31,5 @@ impl fmt::Display for BackendError {
             details: _details,
         } = self;
         write!(f, "{}", message)
-    }
-}
-
-impl BackendError {}
-
-#[cfg(uniffi)]
-pub use uniffimod::{FFIResult, ToUniFFIResult};
-
-#[cfg(uniffi)]
-pub mod uniffimod {
-    use std::sync::Arc;
-
-    use super::BackendError;
-
-    pub type FFIResult<T> = Result<T, Arc<BackendError>>;
-
-    pub trait ToUniFFIResult<T> {
-        fn uniffi(self) -> FFIResult<T>;
-    }
-
-    impl<T> ToUniFFIResult<T> for Result<T, anyhow::Error> {
-        fn uniffi(self) -> FFIResult<T> {
-            self.map_err(|e| Arc::new(BackendError::from(e)))
-        }
-    }
-
-    #[uniffi::export]
-    impl BackendError {
-        fn get_message(&self) -> String {
-            self.message.to_string()
-        }
-
-        fn get_details(&self) -> Vec<String> {
-            self.details.to_owned()
-        }
-
-        fn json(&self) -> String {
-            serde_json::to_string(&self).unwrap()
-        }
     }
 }
