+++ crates/yomikiri-backend/src/utils.rs
@@ -1,6 +1,3 @@
-#[cfg(uniffi)]
-use std::time::{SystemTime, UNIX_EPOCH};
-
 #[cfg(wasm)]
 pub fn set_panic_hook() {
     console_error_panic_hook::set_once();
@@ -21,14 +18,4 @@ pub fn setup_logger() {
 pub(crate) fn time_now() -> f64 {
     let window = web_sys::window().unwrap();
     window.performance().unwrap().now()
-}
-
-#[cfg(uniffi)]
-#[allow(dead_code)]
-pub(crate) fn time_now() -> f64 {
-    let micro = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .unwrap()
-        .as_micros();
-    (micro as f64) / 1000.0
 }
