+++ safari/YomikiriTokenizer/YomikiriTokenizer/YomikiriTokenizer.h
@@ -15,5 +15,4 @@ FOUNDATION_EXPORT const unsigned char YomikiriTokenizerVersionString[];
 
 // In this header, you should import all the public headers of your framework using statements like #import <YomikiriTokenizer/PublicHeader.h>
 
-#include <YomikiriTokenizer/yomikiri_rsFFI.h>
-#include <YomikiriTokenizer/yomikiri_dictionaryFFI.h>
+#include <YomikiriTokenizer/yomikiri_backend_uniffiFFI.h>
